sstables: split stats_metadata into its own header
sstables/types.hh bundled ~15 unrelated on-disk types together. Only
stats_metadata actually needs utils/estimated_histogram.hh (782 lines)
and utils/streaming_histogram.hh by value, plus locator/host_id.hh -
the other metadata structs (validation_metadata, compaction_metadata,
serialization_header) and everything else in the file are unaffected.
Of the 18 users of stats_metadata project-wide, only a handful touch
it directly; the rest were paying for the histogram headers just by
including sstables/types.hh for unrelated declarations.

Move stats_metadata into a new sstables/metadata.hh (which includes
sstables/types.hh for its metadata_base<T> base). Add the histogram
includes directly to the couple of headers (sstables/writer.hh,
sstables/metadata_collector.hh) that were relying on getting them
transitively through sstables/types.hh, and add the new header to
sstables/sstables.hh and metadata_collector.hh where stats_metadata is
used directly.

Targets full-build time via reduced transitive fan-out from
sstables/types.hh (302 direct includers).
